backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-decouple
sqlalchemy
psycopg2-binary
alembic
openai
langchain
langchain-openai
chromadb
supabase
pydantic
pydantic-settings
httpx
redis
celery
pandas
numpy
scikit-learn
matplotlib
seaborn
pytest
pytest-asyncio
black
isort
mypy
